requirements.txt
streamlit
pandas
scikit-learn
joblib
google-generativeai